refactor: split TrackedChangeGroups into presentational components
The 860-line monolith becomes a component folder (DocxToolbar precedent):
index.tsx keeps all state and editor wiring; BookmarkTab, RailHead,
GroupCard and ChangeChip are purely presentational, with shared palette
tokens in styles.ts and the view shapes in types.ts. Verified
render-identical against the monolith (DOM + resolved emotion styles).
